DeviceInfo refactor, do not crash on invalid inputs
* Move the implementation to its own module
* Add ip_address and token properties
* Add tests
